+++ state/bonsai.go
@@ -54,8 +54,8 @@ func (b *Bonsai) GetWithProof(key []byte) ([]byte, iavl.KeyProof, error) {
 	return b.Tree.GetWithProof(key)
 }
 
-func (b *Bonsai) GetVersionedWithProof(key []byte, version uint64) ([]byte, iavl.KeyProof, error) {
-	return b.Tree.GetVersionedWithProof(key, version)
+func (b *Bonsai) GetVersionedWithProof(key []byte, version int64) ([]byte, iavl.KeyProof, error) {
+	return b.Tree.GetVersionedWithProof(key, uint64(version))
 }
 
 func (b *Bonsai) List(start, end []byte, limit int) []Model {
